refactor(MenuAcceptanceTest, MenuGroupAcceptanceTest): 인수테스트 패키지 이동
